+++ emr-serverless-spark-20230808/src/Models/ListSqlStatementContentsRequest.php
@@ -1,0 +1,76 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Emrserverlessspark\V20230808\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class ListSqlStatementContentsRequest extends Model
+{
+    /**
+     * @var string
+     */
+    public $fileName;
+
+    /**
+     * @var int
+     */
+    public $maxResults;
+
+    /**
+     * @var string
+     */
+    public $nextToken;
+    protected $_name = [
+        'fileName' => 'fileName',
+        'maxResults' => 'maxResults',
+        'nextToken' => 'nextToken',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->fileName) {
+            $res['fileName'] = $this->fileName;
+        }
+
+        if (null !== $this->maxResults) {
+            $res['maxResults'] = $this->maxResults;
+        }
+
+        if (null !== $this->nextToken) {
+            $res['nextToken'] = $this->nextToken;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['fileName'])) {
+            $model->fileName = $map['fileName'];
+        }
+
+        if (isset($map['maxResults'])) {
+            $model->maxResults = $map['maxResults'];
+        }
+
+        if (isset($map['nextToken'])) {
+            $model->nextToken = $map['nextToken'];
+        }
+
+        return $model;
+    }
+}