+++ emr-serverless-spark-20230808/src/Models/ListSqlStatementContentsResponseBody/sqlStatementContents.php
@@ -1,0 +1,90 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Emrserverlessspark\V20230808\Models\ListSqlStatementContentsResponseBody;
+
+use AlibabaCloud\Dara\Model;
+
+class sqlStatementContents extends Model
+{
+    /**
+     * @var string
+     */
+    public $contents;
+
+    /**
+     * @var int
+     */
+    public $maxResults;
+
+    /**
+     * @var string
+     */
+    public $nextToken;
+
+    /**
+     * @var int
+     */
+    public $totalCount;
+    protected $_name = [
+        'contents' => 'contents',
+        'maxResults' => 'maxResults',
+        'nextToken' => 'nextToken',
+        'totalCount' => 'totalCount',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->contents) {
+            $res['contents'] = $this->contents;
+        }
+
+        if (null !== $this->maxResults) {
+            $res['maxResults'] = $this->maxResults;
+        }
+
+        if (null !== $this->nextToken) {
+            $res['nextToken'] = $this->nextToken;
+        }
+
+        if (null !== $this->totalCount) {
+            $res['totalCount'] = $this->totalCount;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['contents'])) {
+            $model->contents = $map['contents'];
+        }
+
+        if (isset($map['maxResults'])) {
+            $model->maxResults = $map['maxResults'];
+        }
+
+        if (isset($map['nextToken'])) {
+            $model->nextToken = $map['nextToken'];
+        }
+
+        if (isset($map['totalCount'])) {
+            $model->totalCount = $map['totalCount'];
+        }
+
+        return $model;
+    }
+}